#1079 fix some purchasing tests/service on Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13106 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/purchasing/src/org/opentaps/purchasing/mrp/OpentapsMrpServices.java
@@ -1129,7 +1129,7 @@ public final class OpentapsMrpServices {
                             // -----------------------------------------------------
                             // The components are also loaded thru the configurator
                             BigDecimal positiveEventQuantity = (eventQuantity.signum() > 0 ? eventQuantity : eventQuantity.negate());
-                            serviceResponse = dispatcher.runSync("getManufacturingComponents", UtilMisc.<String, Object>toMap("productId", product.getString("productId"), "quantity", positiveEventQuantity.doubleValue(), "excludeWIPs", Boolean.FALSE, "userLogin", userLogin));
+                            serviceResponse = dispatcher.runSync("getManufacturingComponents", UtilMisc.<String, Object>toMap("productId", product.getString("productId"), "quantity", positiveEventQuantity, "excludeWIPs", Boolean.FALSE, "userLogin", userLogin));
 
                             // Lack of workEffortId signify some routings for the product and
                             // with min/max quantity exist but don't match requested quantity.
@@ -1201,7 +1201,7 @@ public final class OpentapsMrpServices {
 
                                 // -----------------------------------------------------
                                 // The components are also loaded thru the configurator
-                                serviceResponse = dispatcher.runSync("getManufacturingComponents", UtilMisc.<String, Object>toMap("productId", product.getString("productId"), "quantity", proposedOrder.getQuantity().doubleValue(), "excludeWIPs", Boolean.FALSE, "userLogin", userLogin));
+                                serviceResponse = dispatcher.runSync("getManufacturingComponents", UtilMisc.<String, Object>toMap("productId", product.getString("productId"), "quantity", proposedOrder.getQuantity(), "excludeWIPs", Boolean.FALSE, "userLogin", userLogin));
                                 components = (List) serviceResponse.get("components");
                                 String routingId = (String) serviceResponse.get("workEffortId");
                                 if (routingId != null) {
